refactor(aapd-1053): move format-address fn to common for web-comment and forms-web-app
