+++ packages/web-comment/src/routes/decided-appeals/controller.js
@@ -1,4 +1,4 @@
-const { formatAddress } = require('#utils/format-address');
+const { formatAddress } = require('@pins/common/src/lib/format-address');
 const { formatDate } = require('#utils/format-date');
 const { sortByCaseDecisionDate } = require('#utils/appeal-sorting');
 const { apiClient } = require('#utils/appeals-api-client');
